Create CoInusuredInputViewModel and moved logic there for coinsured
