+++ Projects/Contracts/Sources/View/CoInsured/AlreadyExistingCoInsured/InsuredPeopleScreen.swift
@@ -180,19 +180,9 @@ struct InsuredPeopleScreen_Previews: PreviewProvider {
 }
 
 class InsuredPeopleNewScreenModel: ObservableObject {
-    @Published var firstName = ""
-    @Published var lastName = ""
 
     @Published var coInsuredAdded: [CoInsuredModel] = []
     @Published var coInsuredDeleted: [CoInsuredModel] = []
-    @Published var noSSN = false
-    @Published var SSNError: Error?
-    @Published var nameFetchedFromSSN: Bool = false
-    @Published var isLoading: Bool = false
-    @Published var showErrorView: Bool = false
-    @Published var enterManually: Bool = false
-    @PresentableStore var store: ContractStore
-    @Inject var octopus: hOctopus
 
     var resetCoInsured: Void {
         coInsuredAdded = []
@@ -212,60 +202,6 @@ class InsuredPeopleNewScreenModel: ObservableObject {
             }
         } else {
             coInsuredDeleted.append(CoInsuredModel(firstName: firstName, lastName: lastName, SSN: personalNumber))
-        }
-    }
-
-    @MainActor
-    func getNameFromSSN(SSN: String) async {
-        withAnimation {
-            self.SSNError = nil
-            self.isLoading = true
-        }
-        do {
-            let data = try await withCheckedThrowingContinuation {
-                (
-                    continuation: CheckedContinuation<
-                        OctopusGraphQL.PersonalInformationQuery.Data.PersonalInformation, Error
-                    >
-                ) -> Void in
-                let SSNInput = OctopusGraphQL.PersonalInformationInput(personalNumber: SSN)
-                self.octopus.client
-                    .fetch(
-                        query: OctopusGraphQL.PersonalInformationQuery(input: SSNInput),
-                        cachePolicy: .fetchIgnoringCacheCompletely
-                    )
-                    .onValue { value in
-                        if let data = value.personalInformation {
-                            continuation.resume(with: .success(data))
-                        }
-                    }
-                    .onError { graphQLError in
-
-                        continuation.resume(throwing: graphQLError)
-                    }
-            }
-            withAnimation {
-                self.firstName = data.firstName
-                self.lastName = data.lastName
-                self.nameFetchedFromSSN = true
-            }
-
-        } catch let exception {
-            if let exception = exception as? GraphQLError {
-                switch exception {
-                case .graphQLError:
-                    self.enterManually = true
-                case .otherError:
-                    self.enterManually = false
-                }
-            }
-            withAnimation {
-                self.SSNError = exception
-                self.showErrorView = true
-            }
-        }
-        withAnimation {
-            self.isLoading = false
         }
     }
 }
